Move test helper classes to a consistent package (com.nextcloud.test)
Avoids having them mixed up with the tests
