+++ app/src/androidTest/java/com/nextcloud/test/ScreenshotTestRunner.java
@@ -20,7 +20,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.nextcloud.client;
+package com.nextcloud.test;
 
 import android.app.Application;
 import android.content.Context;
