+++ app/src/androidTest/java/com/owncloud/android/ui/dialog/SendShareDialogTest.kt
@@ -24,7 +24,7 @@ package com.owncloud.android.ui.dialog
 import androidx.fragment.app.FragmentManager
 import androidx.test.espresso.intent.rule.IntentsTestRule
 import androidx.test.platform.app.InstrumentationRegistry
-import com.nextcloud.client.TestActivity
+import com.nextcloud.test.TestActivity
 import com.owncloud.android.AbstractIT
 import com.owncloud.android.datamodel.OCFile
 import com.owncloud.android.lib.resources.status.OCCapability
