+++ app/src/androidTest/java/com/owncloud/android/ui/helpers/UriUploaderIT.kt
@@ -2,7 +2,7 @@ package com.owncloud.android.ui.helpers
 
 import android.net.Uri
 import androidx.test.core.app.launchActivity
-import com.nextcloud.client.TestActivity
+import com.nextcloud.test.TestActivity
 import com.owncloud.android.AbstractIT
 import com.owncloud.android.files.services.FileUploader
 import org.junit.Assert
