+++ app/src/debug/java/com/nextcloud/test/TestActivity.kt
@@ -19,7 +19,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.nextcloud.client
+package com.nextcloud.test
 
 import android.os.Bundle
 import android.view.View
